CRUD for comment

+++ src/controllers/task.controller.js
@@ -67,6 +67,10 @@ exports.updateTask = async (req, res) => {
 
 
 exports.deleteTask = async (req, res) => {
+  if(!Number.isInteger(req.params.taskId))
+  {
+    return res.status(400).json({ message: "Invalid TaskID" });
+  }
   try {
     await taskService.deleteTask({
       taskId: Number(req.params.taskId),
